+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Release Notes
 
+## [4.0.5]
+
+### Added
+
+- Added a "Don't show again" option for mainpath warnings.
+
+### Fixed
+
+- Fixed enum struct attributes missing.
+
 ## [4.0.4]
 
 ### Fixed